requirements.txt
dask[dataframe]
azure-functions
azure-storage-blob
spatialite 
pysqlite3
pandas
pyarrow